Refactors and adds comments to clr.py to explain the improvements
