+++ app/grants/management/commands/analytics_clr.py
@@ -22,13 +22,13 @@ from django.core.management.base import BaseCommand
 from django.db import connection
 from django.utils import timezone
 
-from grants.clr import calculate_clr, fetch_grants, get_summed_contribs_query, get_totals_by_pair, normalise
+from grants.clr import calculate_clr, get_totals_by_pair, normalise
+from grants.clr_data_src import fetch_grants, fetch_summed_contributions
 from grants.models import GrantCLR
 
 
 def analytics_clr(from_date=None, clr_round=None, network='mainnet'):
     # setup
-    # clr_calc_start_time = timezone.now()
     debug_output = [['grant_id', 'grant_title', 'number_contributions', 'contribution_amount', 'clr_amount']]
 
     # one-time data call
@@ -37,38 +37,22 @@ def analytics_clr(from_date=None, clr_round=None, network='mainnet'):
 
     print(total_pot)
 
+    # fetch data
     grants = fetch_grants(clr_round, network)
+    curr_agg, trust_dict = fetch_summed_contributions(grants, clr_round, network)
 
-    # collect contributions for clr_round into temp table
-    initial_query = get_summed_contribs_query(grants, clr_round.start_date, clr_round.end_date, clr_round.contribution_multiplier, network)
-
-    # open cursor and execute the groupBy sum for the round
-    with connection.cursor() as cursor:
-        curr_agg = {}
-        trust_dict = {}
-        # execute to populate shared state for the round
-        cursor.execute(initial_query) # (we could potential do better here by sharing this temp table between rounds)
-        for _row in cursor.fetchall():
-            if not curr_agg.get(_row[0]):
-                curr_agg[_row[0]] = {}
-
-            trust_dict[_row[1]] = _row[3]
-            curr_agg[_row[0]][_row[1]] = _row[2]
-
-        ptots = get_totals_by_pair(curr_agg)
-        bigtot, totals = calculate_clr(curr_agg, ptots, trust_dict, v_threshold, total_pot)
-
-        # normalise against a deepcopy of the totals to avoid mutations
-        curr_grants_clr = normalise(bigtot, totals, total_pot)
-        
-        # calculate clr analytics output
-        for grant in grants:
-            num_contribs, contrib_amount, clr_amount = curr_grants_clr.get(grant.id, {'num_contribs': 0, 'contrib_amount': 0, 'clr_amount': None}).values()
-            # debug_output.append([grant.id, grant.title, num_contribs, contrib_amount, clr_amount])
-            debug_output.append([grant.id, grant.title, grant.positive_round_contributor_count, float(grant.amount_received_in_round), clr_amount])
+    # run calculation
+    ptots = get_totals_by_pair(curr_agg)
+    bigtot, totals = calculate_clr(curr_agg, ptots, trust_dict, v_threshold, total_pot)
+    curr_grants_clr = normalise(bigtot, totals, total_pot)
+    
+    # calculate clr analytics output
+    for grant in grants:
+        num_contribs, contrib_amount, clr_amount = curr_grants_clr.get(grant.id, {'num_contribs': 0, 'contrib_amount': 0, 'clr_amount': None}).values()
+        # debug_output.append([grant.id, grant.title, num_contribs, contrib_amount, clr_amount])
+        debug_output.append([grant.id, grant.title, grant.positive_round_contributor_count, float(grant.amount_received_in_round), clr_amount])
 
     return debug_output
-
 
 
 class Command(BaseCommand):
