+++ app/grants/views.py
@@ -69,7 +69,7 @@ from dashboard.utils import get_web3
 from economy.models import Token as FTokens
 from economy.utils import convert_token_to_usdt
 from eth_account.messages import defunct_hash_message
-from grants.clr_old import fetch_data
+from grants.clr_data_src import fetch_contributions
 from grants.models import (
     CartActivity, Contribution, Flag, Grant, GrantAPIKey, GrantBrandingRoutingPolicy, GrantCLR, GrantCollection,
     GrantTag, GrantType, MatchPledge, Subscription,
@@ -3587,7 +3587,7 @@ def get_clr_sybil_input(request, round_id):
         limit = data['limit'] if data['limit'] else 100
 
         # fetch grant contributions needed for round
-        __, all_clr_contributions = fetch_data(clr)
+        all_clr_contributions = fetch_contributions(clr)
         total_count = all_clr_contributions.count()
 
         # extract only needed fields
